WIP - intitial support for recording all pvs at all depths.
Not quite ready yet. Needs backwards compatibility for multipv=1,
We should also consider a full backwards compatibility and dropping the
depth key completely when multipv is excluded from the job description
so we don't waste bandwidth for info we don't store. Also, I need to
find out if my changes to the info parsing will have adverse effects on
scenarios I'm not familiar with. Also need to find out if we can depend
on stockfish reporting depths an multipvs in order. And finally need to
determine if I can make the json less fat by de-duplicating some of the
values that are shared amongst all pvs at a certain depth.